+++ crates/nostr/src/nips/nip21.rs
@@ -13,8 +13,10 @@ use core::str::FromStr;
 
 use bech32::Fe32;
 
-use super::nip01::Coordinate;
-use super::nip19::{self, FromBech32, Nip19, Nip19Event, Nip19Prefix, Nip19Profile, ToBech32};
+use super::nip19::{
+    self, FromBech32, Nip19, Nip19Coordinate, Nip19Event, Nip19Prefix, Nip19Profile, ToBech32,
+};
+use crate::nips::nip01::Coordinate;
 use crate::{EventId, PublicKey};
 
 /// URI scheme
@@ -101,6 +103,7 @@ impl NostrURI for PublicKey {}
 impl NostrURI for EventId {}
 impl NostrURI for Nip19Profile {}
 impl NostrURI for Nip19Event {}
+impl NostrURI for Nip19Coordinate {}
 impl NostrURI for Coordinate {}
 
 /// A representation any `NIP21` object. Useful for decoding
@@ -117,7 +120,7 @@ pub enum Nip21 {
     /// nostr::nevent
     Event(Nip19Event),
     /// nostr::naddr
-    Coordinate(Coordinate),
+    Coordinate(Nip19Coordinate),
 }
 
 impl From<Nip21> for Nip19 {